(PC-37176)[PRO] chore: remove unused code
