+++ pro/src/components/IndividualOffer/StocksEventCreation/StocksCalendar/StocksCalendarFilters/StocksCalendarFilters.tsx
@@ -1,7 +1,7 @@
 import { StocksOrderedBy } from 'apiClient/v1'
 import { PriceCategoryResponseModel } from 'apiClient/v1/models/PriceCategoryResponseModel'
 import { OFFER_WIZARD_MODE } from 'commons/core/Offers/constants'
-import { getPriceCategoryOptions } from 'components/IndividualOffer/StocksEventEdition/getPriceCategoryOptions'
+import { getPriceCategoryOptions } from 'components/IndividualOffer/PriceCategoriesScreen/form/getPriceCategoryOptions'
 import fullRefreshIcon from 'icons/full-refresh.svg'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
