--- pro/src/components/IndividualOffer/StocksEventCreation/StocksEventCreation.module.scss
@@ -1,11 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-
-.help-section {
-  margin: rem.torem(56px) 0 rem.torem(24px);
-  width: 100%;
-}
-
-.light-banner {
-  width: 100%;
-  margin-top: rem.torem(32px);
-}
