--- pro/src/components/IndividualOffer/StocksEventEdition/StocksEventEdition.module.scss
@@ -1,99 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_a11y.scss" as a11y;
-
-.add-dates-button {
-  display: flex;
-  justify-content: center;
-  margin-bottom: rem.torem(32px);
-}
-
-.stock-table-container {
-  overflow-x: auto;
-}
-
-.stock-table {
-  margin-top: rem.torem(16px);
-  table-layout: fixed;
-  width: 100%;
-}
-
-.table-head {
-  @include fonts.body-accent-xs;
-
-  text-align: left;
-  padding-bottom: rem.torem(16px);
-  padding-right: rem.torem(16px);
-}
-
-// Cancel additional padding from the filter results row
-.filters-active .table-head {
-  padding-bottom: 0;
-}
-
-.header-name {
-  @include fonts.body-accent-xs;
-}
-
-.filter-input {
-  display: flex;
-  gap: rem.torem(8px);
-  margin-bottom: rem.torem(24px);
-}
-
-.table-body {
-  padding-bottom: rem.torem(16px);
-}
-
-.table-row {
-  vertical-align: bottom;
-}
-
-.head-date {
-  width: rem.torem(110px);
-}
-
-.head-time {
-  width: rem.torem(90px);
-}
-
-.head-price {
-  width: rem.torem(171px);
-}
-
-.head-booking-limit-datetime {
-  width: rem.torem(110px);
-
-  // force header to break line
-  padding-right: rem.torem(4px);
-}
-
-.head-remaining-quantity {
-  width: rem.torem(180px);
-}
-
-.head-booking-quantity {
-  white-space: nowrap;
-  width: rem.torem(100px);
-}
-
-.head-actions {
-  width: rem.torem(40px);
-}
-
-.data {
-  padding-right: rem.torem(12px);
-  padding-bottom: rem.torem(8px);
-}
-
-.style-bookings-quantity {
-  padding-bottom: rem.torem(16px);
-}
-
-.quantity-input {
-  width: 10rem;
-}
-
-.visually-hidden {
-  @include a11y.visually-hidden;
-}
