+++ pro/src/components/IndividualOffer/StocksEventEdition/types.ts
