--- pro/src/components/StocksEventList/StocksEventList.module.scss
@@ -1,200 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/variables/_z-index.scss" as zIndex;
-@use "styles/mixins/_size.scss" as size;
-@use "styles/mixins/_a11y.scss" as a11y;
-
-.stock-event-table {
-  width: 100%;
-  margin-top: rem.torem(16px);
-  margin-bottom: rem.torem(32px);
-  display: flex;
-  flex-direction: column;
-}
-
-.select-all-container {
-  display: flex;
-  gap: rem.torem(16px); 
-  align-items: center;
-  margin-top: rem.torem(24px); 
-}
-
-.header {
-  vertical-align: bottom;
-  text-align: left;
-  padding-bottom: rem.torem(16px);
-}
-
-.row-head {
-  display: none;
-  flex-direction: column;
-}
-
-.header-name {
-  @include fonts.body-accent-xs;
-}
-
-.row {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: rem.torem(16px);
-}
-
-.data {
-  display: flex;
-  justify-content: space-between;
-  padding: rem.torem(12px) rem.torem(8px);
-  align-items: center;
-
-  &:last-child {
-    text-align: center;
-  }
-}
-
-.data::before {
-  @include fonts.body-accent-xs;
-
-  content: attr(data-label) " :";
-  text-align: left;
-  margin-right: rem.torem(16px);
-}
-
-.body .row:nth-child(n + 1):nth-child(odd) {
-  background-color: var(--color-background-subtle);
-}
-
-.date-cell-wrapper {
-  display: flex;
-  flex-direction: row;
-  gap: rem.torem(8px);
-}
-
-.capitalize {
-  text-transform: capitalize;
-}
-
-.actions {
-  z-index: zIndex.$recurrence-sticky-bar;
-}
-
-.visually-hidden {
-  @include a11y.visually-hidden;
-}
-
-.filter-input {
-  margin-bottom: rem.torem(16px);
-}
-
-.recurrences-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 rem.torem(24px);
-}
-
-@media (min-width: size.$tablet) {
-  .recurrences-button {
-    margin: rem.torem(32px) auto rem.torem(32px) auto;
-  }
-
-  .stock-event-table {
-    margin-top: rem.torem(16px);
-    margin-bottom: rem.torem(32px);
-    display: table;
-  }
-
-  .body .row:nth-child(n + 1):nth-child(odd) {
-    background-color: var(--color-background-default);
-  }
-
-  .row-head {
-    display: table-row;
-    padding: rem.torem(16px);
-    background-color: var(--color-background-subtle);
-  }
-
-  .row {
-    display: table-row;
-    padding: rem.torem(16px);
-  }
-
-  .header {
-    @include fonts.body-accent-xs;
-
-    cursor: pointer;
-    padding:  rem.torem(8px) rem.torem(8px) rem.torem(8px) 0;
-
-    &:first-child {
-      padding-left: rem.torem(8px);
-      border-top-left-radius: rem.torem(8px);
-    }
-
-    &:last-child {
-      border-top-right-radius: rem.torem(8px);
-      padding-right: rem.torem(8px);
-    }
-  }
-
-  .data {
-    display: table-cell;
-    padding: rem.torem(12px) 0;
-
-    &::before {
-      content: none;
-      display: none;
-    }
-  }
-
-  .row:not(:last-child) {
-    border-bottom: 1px solid var(--color-border-subtle);
-  }
-
-  $icon-size: rem.torem(20px);
-
-  .clear-icon {
-    svg {
-      width: $icon-size;
-      fill: var(--color-icon-default);
-    }
-  }
-
-  .checkbox-column {
-    width: rem.torem(32px);
-    padding-right: rem.torem(8px);
-    text-transform: capitalize;
-  }
-
-  .date-column {
-    min-width: rem.torem(192px);
-  }
-
-  .price-column {
-    min-width: rem.torem(200px);
-  }
-
-  .filter-input {
-    display: flex;
-    gap: rem.torem(8px);
-    margin-bottom: rem.torem(24px);
-  }
-
-  .time-column {
-    min-width: rem.torem(88px);
-  }
-
-  .booking-limit-date-column {
-    min-width: rem.torem(200px);
-  }
-
-  .quantity-column {
-    min-width: rem.torem(100px);
-  }
-
-  .bookings-quantity-column {
-    min-width: rem.torem(120px);
-  }
-
-  .actions-column {
-    min-width: $icon-size;
-  }
-}
