+++ pro/src/pages/IndividualOffer/components/ActionBar/ActionBar.tsx
@@ -5,7 +5,6 @@ import {
   OFFER_WIZARD_MODE,
 } from 'commons/core/Offers/constants'
 import { computeIndividualOffersUrl } from 'commons/core/Offers/utils/computeIndividualOffersUrl'
-import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useNotification } from 'commons/hooks/useNotification'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { ActionsBarSticky } from 'components/ActionsBarSticky/ActionsBarSticky'
@@ -43,9 +42,6 @@ export const ActionBar = ({
   const mode = useOfferWizardMode()
   const backOfferUrl = computeIndividualOffersUrl({})
   const notify = useNotification()
-  const isEventWithOpeningHoursEnabled = useActiveFeature(
-    'WIP_ENABLE_EVENT_WITH_OPENING_HOUR'
-  )
 
   const Left = (): JSX.Element => {
     if (mode === OFFER_WIZARD_MODE.CREATION) {
@@ -63,8 +59,7 @@ export const ActionBar = ({
 
     if (
       mode === OFFER_WIZARD_MODE.EDITION &&
-      step === INDIVIDUAL_OFFER_WIZARD_STEP_IDS.STOCKS &&
-      isEventWithOpeningHoursEnabled
+      step === INDIVIDUAL_OFFER_WIZARD_STEP_IDS.STOCKS
     ) {
       return (
         <Button onClick={onClickPrevious} variant={ButtonVariant.SECONDARY}>
