+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummaryStocks/IndividualOfferSummaryStocks.tsx
@@ -2,7 +2,6 @@
 
 import { useIndividualOfferContext } from 'commons/context/IndividualOfferContext/IndividualOfferContext'
 import { INDIVIDUAL_OFFER_WIZARD_STEP_IDS } from 'commons/core/Offers/constants'
-import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { IndividualOfferLayout } from 'components/IndividualOfferLayout/IndividualOfferLayout'
 import { ActionBar } from 'pages/IndividualOffer/components/ActionBar/ActionBar'
@@ -15,17 +14,13 @@ const IndividualOfferSummaryStocks = (): JSX.Element | null => {
   const mode = useOfferWizardMode()
   const { offer } = useIndividualOfferContext()
 
-  const isEventWithOpeningHoursEnabled = useActiveFeature(
-    'WIP_ENABLE_EVENT_WITH_OPENING_HOUR'
-  )
-
   if (offer === null) {
     return <Spinner />
   }
 
   return (
     <IndividualOfferLayout title="Récapitulatif" offer={offer} mode={mode}>
-      {isEventWithOpeningHoursEnabled && offer.isEvent ? (
+      {offer.isEvent ? (
         <IndividualOfferSummaryStocksCalendarScreen offer={offer} />
       ) : (
         <IndividualOfferSummaryStocksScreen />
