+++ pro/src/pages/IndividualOfferWizard/Stocks/Stocks.tsx
@@ -1,10 +1,6 @@
 import { useIndividualOfferContext } from 'commons/context/IndividualOfferContext/IndividualOfferContext'
-import { OFFER_WIZARD_MODE } from 'commons/core/Offers/constants'
-import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { StocksCalendar } from 'components/IndividualOffer/StocksEventCreation/StocksCalendar/StocksCalendar'
-import { StocksEventCreation } from 'components/IndividualOffer/StocksEventCreation/StocksEventCreation'
-import { StocksEventEdition } from 'components/IndividualOffer/StocksEventEdition/StocksEventEdition'
 import { StocksThing } from 'components/IndividualOffer/StocksThing/StocksThing'
 import { IndividualOfferLayout } from 'components/IndividualOfferLayout/IndividualOfferLayout'
 import { getTitle } from 'components/IndividualOfferLayout/utils/getTitle'
@@ -13,10 +9,6 @@ import { Spinner } from 'ui-kit/Spinner/Spinner'
 export const Stocks = (): JSX.Element | null => {
   const { offer, publishedOfferWithSameEAN } = useIndividualOfferContext()
   const mode = useOfferWizardMode()
-
-  const isEventWithOpeningHoursEnabled = useActiveFeature(
-    'WIP_ENABLE_EVENT_WITH_OPENING_HOUR'
-  )
 
   // Here we display a spinner because when the router transitions from
   // Informations form to Stocks form the setOffer after the submit is not
@@ -32,15 +24,7 @@ export const Stocks = (): JSX.Element | null => {
       return <StocksThing offer={offer} />
     }
 
-    if (isEventWithOpeningHoursEnabled) {
-      return <StocksCalendar offer={offer} mode={mode} />
-    }
-
-    return mode === OFFER_WIZARD_MODE.CREATION ? (
-      <StocksEventCreation offer={offer} />
-    ) : (
-      <StocksEventEdition offer={offer} />
-    )
+    return <StocksCalendar offer={offer} mode={mode} />
   }
 
   return (
